requirements.txt
attrs==21.4.0
httpx==0.23.0
aiogram==2.19.0
sentry-sdk==1.45.1
beautifulsoup4==4.11.1
python-decouple==3.7
pyinstaller==6.0.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
